python/cuda_parallel/pyproject.toml
# Copyright (c) 2025, NVIDIA CORPORATION & AFFILIATES. ALL RIGHTS RESERVED.
#
# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception

[build-system]
requires = ["scikit-build-core>=0.10"]
build-backend = "scikit_build_core.build"

[project]
name = "cuda-parallel"
description = "Experimental Core Library for CUDA Python"
authors = [{ name = "NVIDIA Corporation" }]
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Environment :: GPU :: NVIDIA CUDA",
  "License :: OSI Approved :: Apache Software License",
]
requires-python = ">=3.9"
dependencies = [
  "cuda-cccl",
  "numba>=0.60.0",
  "cuda-python==12.*",
  "cupy-cuda12x",
]
dynamic = ["version"]
readme = { file = "README.md", content-type = "text/markdown" }

[project.optional-dependencies]
test = ["pytest", "pytest-xdist", "cupy-cuda12x", "typing_extensions"]

[project.urls]
Homepage = "https://developer.nvidia.com/"

[tool.scikit-build]
minimum-version = "build-system.requires"
build-dir = "build/{wheel_tag}"

[tool.scikit-build.cmake]
version = ">=3.21"
args = []
build-type = "Release"
source-dir = "."

[tool.scikit-build.ninja]
version = ">=1.11"
make-fallback = true

[tool.scikit-build.wheel]
py-api = "py3"

[tool.scikit-build.wheel.packages]
"cuda" = "cuda"
"cuda/parallel" = "cuda/parallel"
"cuda/parallel/experimental" = "cuda/parallel/experimental"

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
input = "cuda/parallel/_version.py"
# use default regex

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = ["numba.*", "llvmlite", "cupy"]
ignore_missing_imports = true
follow_imports = "skip"

[tool.ruff]
extend = "../../pyproject.toml"

[tool.ruff.lint.isort]
known-first-party = ["cuda.parallel"]
